Move catalog config contributor to core
