+++ extensions/semantic-models/src/main/java/org/apache/polaris/service/catalog/semanticmodel/SemanticModelConfigEndpoints.java
@@ -26,7 +26,7 @@ import java.util.Set;
 import org.apache.iceberg.rest.Endpoint;
 import org.apache.polaris.core.config.FeatureConfiguration;
 import org.apache.polaris.core.config.RealmConfig;
-import org.apache.polaris.service.spi.CatalogConfigEndpointContributor;
+import org.apache.polaris.core.rest.CatalogConfigEndpointContributor;
 
 @ApplicationScoped
 @Priority(500)
