+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/iceberg/IcebergRestConfigEndpoints.java
@@ -23,7 +23,7 @@ import jakarta.annotation.Priority;
 import jakarta.enterprise.context.ApplicationScoped;
 import java.util.Set;
 import org.apache.iceberg.rest.Endpoint;
-import org.apache.polaris.service.spi.CatalogConfigEndpointContributor;
+import org.apache.polaris.core.rest.CatalogConfigEndpointContributor;
 
 @ApplicationScoped
 @Priority(100)
